refactor: refactor names and folder structure
