+++ src/components/Loading.tsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import { Component } from 'react';
 
 export default class Loading extends Component {
   render() {
