+++ src/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App';
 import './index.css';
-import ErrorBoundary from './error';
+import ErrorBoundary from './components/Error';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
